arch/pyproject.toml
[project]
name = "arch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "camelot-py>=1.0.0",
    "ghostscript>=0.7",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "tabula-py>=2.10.0",
]
